add base files work on detail screen in progress

+++ app/src/main/java/zw/co/nm/moviedb/models/network/GetMovieDetailResponse.kt
@@ -37,7 +37,7 @@ data class GetMovieDetailResponse(
     @SerializedName("release_date")
     val releaseDate: String,
     @SerializedName("revenue")
-    val revenue: Int,
+    val revenue: Long,
     @SerializedName("runtime")
     val runtime: Int,
     @SerializedName("spoken_languages")
